#161: Reduce the number of copies in code (#218)
# Description

This commit changes many functions from returning an `std::string` to
returning a `const std::string &`. It also attempts to cache more
strings/vectors.

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)

# How Has This Been Tested?

Tested on macOS 15.7.3.

# Checklist:

- [x] My code follows the style guidelines of this project (`black` for
Python
  code, `.clang-format` in the `src/jyut-dict` directory for C++)
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have translated my user-facing strings to all
currently-supported languages
- [x] I have made corresponding changes to the documentation

+++ src/jyut-dict/logic/entry/test/TestEntry/tst_entry.cpp
@@ -142,7 +142,7 @@ void TestEntry::cantoneseToneNumbers()
     Entry entry{"", "", jyutping, "", {}};
 
     auto result = entry.getJyutpingNumbers();
-    std::vector<int> expected{4, 1, 2};
+    std::vector<uint8_t> expected{4, 1, 2};
     QCOMPARE(result, expected);
 
     entry.setJyutping("mei5 gwok3");
@@ -196,7 +196,7 @@ void TestEntry::mandarinToneNumbers()
     Entry entry{"", "", "", pinyin, {}};
 
     auto result = entry.getPinyinNumbers();
-    std::vector<int> expected{2, 1, 4};
+    std::vector<uint8_t> expected{2, 1, 4};
     QCOMPARE(result, expected);
 
     entry.setPinyin("mei3 guo2");
